requirements.txt
python-dotenv
tensorflow
pillow
openai
transformers
ipython
langchain